+++ tests/graphs/task1/expected.dot
@@ -1,0 +1,21 @@
+digraph  {
+1;
+2;
+3;
+0;
+4;
+5;
+6;
+7;
+8;
+1 -> 2 [key=0, label=a];
+2 -> 3 [key=0, label=a];
+3 -> 0 [key=0, label=a];
+0 -> 1 [key=0, label=a];
+0 -> 4 [key=0, label=b];
+4 -> 5 [key=0, label=b];
+5 -> 6 [key=0, label=b];
+6 -> 7 [key=0, label=b];
+7 -> 8 [key=0, label=b];
+8 -> 0 [key=0, label=b];
+}
