+++ tests/test_task1.py
@@ -1,17 +1,21 @@
 import filecmp
 
-import project.graph_lib as gl
+from project.task1 import (
+    get_graph_by_name,
+    get_graph_info,
+    make_labeled_two_cycles_graph,
+)
 
 
 def test_load_graph_from_name():
-    graph = gl.get_graph_by_name("skos")
+    graph = get_graph_by_name("skos")
     assert graph.number_of_nodes() == 144
     assert graph.size() == 252
 
 
 def test_get_graph_info():
-    graph = gl.get_graph_by_name("skos")
-    nodes, edges, labels = gl.get_graph_info(graph)
+    graph = get_graph_by_name("skos")
+    nodes, edges, labels = get_graph_info(graph)
 
     assert nodes == graph.number_of_nodes()
     assert edges == graph.size()
@@ -20,9 +24,9 @@ def test_get_graph_info():
 
 
 def test_make_labeled_two_cycles_graph():
-    path = "./tests/graphs/task_1/actual.dot"
-    expected = "./tests/graphs/task_1/expected.dot"
+    path = "./tests/graphs/task1/actual.dot"
+    expected = "./tests/graphs/task1/expected.dot"
 
-    gl.make_labeled_two_cycles_graph(3, 5, ("a", "b"), path)
+    make_labeled_two_cycles_graph(3, 5, ("a", "b"), path)
 
     assert filecmp.cmp(path, expected, shallow=False)
